feat(sentry): remove sentry env support
Since we are going to do trunk based dev on all platforms, we can safely remove the Sentry environment value. Each error reported to Sentry could either happen on a prod or acc environment, but that isn't relevant. Since it may as well happen on prod since the same build is deployed there.